+++ lib/core/utils/pollyfill-elements-from-point.js
